+++ lib/index.ts
@@ -1,5 +1,4 @@
 export * from './types';
-export * from './builder';
 export * from './writer';
 export * from './errors';
 export * from './utils';
@@ -10,3 +9,8 @@ export * from './sql/operators';
 export * from './sql/functions/type';
 export * from './sql/functions/time';
 export * from './sql/functions/rand';
+
+export * from './impl/base';
+export * from './impl/stateful';
+export * from './impl/stateless';
+export * from './impl/wrapper';